requirements.txt
Flask>=2.0
SQLAlchemy>=1.4
python-dotenv>=0.19
requests>=2.25     # For making API calls (WhatsApp, potentially OpenAI media download)
openai>=1.3        # OpenAI library for Whisper and GPT
psycopg2-binary    # PostgreSQL driver for SQLAlchemy